# server/openapi.yaml
openapi: 3.0.0
info:
  title: Markdown Ticket API
  version: 1.0.0
  description: |-
    REST API for Markdown Ticket - an AI-powered Kanban board where tickets are markdown files.

    ## Key Features
    - **Multi-project support**: Manage CRs across multiple projects
    - **File-based storage**: All CRs stored as markdown files with YAML frontmatter
    - **Real-time updates**: SSE endpoint for live change notifications
    - **Git integration**: Version control for all ticket changes

    ## Authentication
    Currently, the API does not require authentication. Future versions may add API key or OAuth support.

    ## Error Handling
    All endpoints return consistent error responses with `error` and `message` fields.
  contact:
    name: Markdown Ticket
    url: 'https://github.com/example/markdown-ticket'
  license:
    name: MIT
    url: 'https://opensource.org/licenses/MIT'
servers:
  - url: 'http://localhost:3001'
    description: Development server
...
components:
  schemas:
...
    Document:
      type: object
      required:
        - path
        - name
      properties:
        path:
          type: string
          description: Relative path from project root
          example: docs/architecture.md
        name:
          type: string
          description: Document filename
          example: architecture.md
        content:
          type: string
          description: Document content (when requested)
    Error400:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Invalid request parameters
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    Error404:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          example: Not Found
        message:
          type: string
          example: Resource not found
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error400'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error404'
...
paths:
...
  /api/documents:
    get:
      summary: Discover documents for a project
      tags:
        - Documents
      parameters:
        - name: projectId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of discovered documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Document'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/documents/content:
    get:
      summary: Get document content
      tags:
        - Documents
      parameters:
        - name: projectId
          in: query
          required: true
          schema:
            type: string
        - name: path
          in: query
          required: true
          schema:
            type: string
          description: Relative path to document
      responses:
        '200':
          description: Document content
          content:
            text/plain:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          description: Access denied or invalid file path
        '404':
          $ref: '#/components/responses/NotFound'
  /api/documents/configure:
    post:
      summary: Configure document paths for a project
      tags:
        - Documents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - projectId
                - documentPaths
              properties:
                projectId:
                  type: string
                documentPaths:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Configuration updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...